+++ flint-spark-integration/src/main/scala/org/opensearch/flint/spark/refresh/AutoIndexRefresher.scala
@@ -1,0 +1,99 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.flint.spark.refresh
+
+import org.opensearch.flint.spark.{FlintSparkIndex, FlintSparkIndexOptions}
+import org.opensearch.flint.spark.FlintSparkIndex.{quotedTableName, StreamingRefresh}
+
+import org.apache.spark.sql.{DataFrame, Row, SparkSession}
+import org.apache.spark.sql.flint.FlintDataSourceV2.FLINT_DATASOURCE
+import org.apache.spark.sql.flint.config.FlintSparkConf
+import org.apache.spark.sql.flint.config.FlintSparkConf.CHECKPOINT_MANDATORY
+import org.apache.spark.sql.streaming.{DataStreamWriter, Trigger}
+
+/**
+ * Index refresher that auto refreshes the index.
+ *
+ * @param indexName
+ *   Flint index name
+ * @param index
+ *   Flint index
+ */
+class AutoIndexRefresher(indexName: String, index: FlintSparkIndex)
+    extends FlintSparkIndexRefresher {
+
+  override def start(spark: SparkSession, flintSparkConf: FlintSparkConf): Option[String] = {
+    val options = index.options
+    val tableName = index.metadata().source
+    index match {
+      // Flint index has specialized logic and capability for incremental refresh
+      case refresh: StreamingRefresh =>
+        logInfo("Start refreshing index in streaming style")
+        val job =
+          refresh
+            .buildStream(spark)
+            .writeStream
+            .queryName(indexName)
+            .format(FLINT_DATASOURCE)
+            .options(flintSparkConf.properties)
+            .addSinkOptions(options, flintSparkConf)
+            .start(indexName)
+        Some(job.id.toString)
+      case _ =>
+        // Otherwise, fall back to foreachBatch + batch refresh
+        logInfo("Start refreshing index in foreach streaming style")
+        val job = spark.readStream
+          .options(options.extraSourceOptions(tableName))
+          .table(quotedTableName(tableName))
+          .writeStream
+          .queryName(indexName)
+          .addSinkOptions(options, flintSparkConf)
+          .foreachBatch { (batchDF: DataFrame, _: Long) =>
+            new FullManualIndexRefresher(indexName, index, Some(batchDF))
+              .start(spark, flintSparkConf)
+            () // discard return value above and return unit to use right overridden method
+          }
+          .start()
+        Some(job.id.toString)
+    }
+  }
+
+  // Using Scala implicit class to avoid breaking method chaining of Spark data frame fluent API
+  private implicit class FlintDataStreamWriter(val dataStream: DataStreamWriter[Row]) {
+
+    def addSinkOptions(
+        options: FlintSparkIndexOptions,
+        flintSparkConf: FlintSparkConf): DataStreamWriter[Row] = {
+      dataStream
+        .addCheckpointLocation(options.checkpointLocation(), flintSparkConf.isCheckpointMandatory)
+        .addRefreshInterval(options.refreshInterval())
+        .addOutputMode(options.outputMode())
+        .options(options.extraSinkOptions())
+    }
+
+    def addCheckpointLocation(
+        checkpointLocation: Option[String],
+        isCheckpointMandatory: Boolean): DataStreamWriter[Row] = {
+      checkpointLocation match {
+        case Some(location) => dataStream.option("checkpointLocation", location)
+        case None if isCheckpointMandatory =>
+          throw new IllegalStateException(
+            s"Checkpoint location is mandatory for incremental refresh if ${CHECKPOINT_MANDATORY.key} enabled")
+        case _ => dataStream
+      }
+    }
+
+    def addRefreshInterval(refreshInterval: Option[String]): DataStreamWriter[Row] = {
+      refreshInterval
+        .map(interval => dataStream.trigger(Trigger.ProcessingTime(interval)))
+        .getOrElse(dataStream)
+    }
+
+    def addOutputMode(outputMode: Option[String]): DataStreamWriter[Row] = {
+      outputMode.map(dataStream.outputMode).getOrElse(dataStream)
+    }
+  }
+}
